+++ package.json
@@ -50,6 +50,7 @@
         "passport-kakao": "^1.0.1",
         "reflect-metadata": "^0.1.14",
         "rxjs": "^7.8.1",
+        "socket.io": "^4.7.4",
         "typeorm": "^0.3.19",
         "typeorm-naming-strategies": "^4.1.0"
     },
