+++ public/styles/common.css
@@ -623,10 +623,15 @@ section {
 
 .chatInput {
     width: 90%;
-    height: 10%;
+    height: 8%;
     justify-content: center;
     text-align: center;
     margin-top: 30px;
+    display: flex;
+    flex-direction: row;
+}
+#chatInputText {
+    width: 10px;
 }
 
 .chatInput > input {
