--- src/chat/chat.interface.ts
@@ -1,20 +1,0 @@
-//export interface IThread {
-//    createdAt: Date;
-//    id: number;
-//    participants: IParticipant[];
-//    updatedAt: Date;
-//}
-
-export interface CreateChatDto {
-    id: number;
-    content: string;
-    createdAt: Date;
-    updatedAt: Date;
-}
-
-//export interface IParticipant {
-//    createdAt: Date;
-//    id: number;
-//    updatedAt: Date;
-//    userId: number;
-//}
