+++ src/chat/intercepter/user.data.intercepter.ts
@@ -6,6 +6,7 @@
 //import { find } from 'lodash';
 //import { Observable } from 'rxjs';
 
+//TODO 추가기능 구현 예정
 //@Injectable()
 //export class WsUserInfo implements NestInterceptor {
 //    constructor(
@@ -13,29 +14,12 @@
 //        private readonly configService: ConfigService,
 //    ) {}
 //    intercept(context: ExecutionContext, next: CallHandler<any>): Observable<any> | Promise<Observable<any>> {
-//        r
+
 //    }
 //    private readonly secretKey = this.configService.get<string>('JWT_SECRET_KEY');
 
 //    canActivate(context: ExecutionContext): any {
 //        const token = context.switchToWs().getClient().handshake.headers.cookie.split('=')[1];
 //        console.log('token', token, context.switchToWs().getClient().handshake.headers);
-
-//        try {
-//            const verifyToken = async (token: string): Promise<any> => {
-//                const verify = jwt.verify(token, process.env.JWT_SECRET_KEY);
-//                console.log('verify', verify, typeof verify.sub);
-//                const userId = verify.sub;
-//                if (!verify) {
-//                    throw new UnauthorizedException('인증이 유효하지 않습니다. ');
-//                } else {
-//                    //next()
-//                    const findUser = await this.userService.findByKakaoIdGetUserName(+userId);
-//                    console.log('findUser', findUser); //findUser User { nickname: '유은지' }
-//                    return (context.switchToWs().getClient().handshake.auth = findUser); //TypeError: Cannot create property 'user' on string 'first'
-//                }
-//            };
-//            verifyToken(token);
-//        } catch (err) {}
 //    }
 //}
