+++ src/chat/schema/chat.schema.ts
@@ -1,6 +1,5 @@
 import { Prop, Schema, SchemaFactory } from '@nestjs/mongoose';
-import { number } from 'joi';
-import { Date, HydratedDocument, ObjectId } from 'mongoose';
+import { HydratedDocument, ObjectId } from 'mongoose';
 import { ObjectIdColumn } from 'typeorm';
 
 export type ChatDocument = HydratedDocument<Chat>;
@@ -21,11 +20,5 @@ export class Chat {
 
     @Prop({ required: true })
     content: string;
-
-    //@Prop({ type: Date, default: () => new Date() })
-    //createdAt: Date;
-
-    //@Prop({ type: Date, default: () => new Date() })
-    //updatedAt: Date;
 }
 export const ChatSchema = SchemaFactory.createForClass(Chat);
